requirements.txt
aiohttp>=3.3.2    
aiosqlite==0.3.0    
numpy>=1.14.5
PyYAML>=3.13     
git+https://github.com/hard-codr/sirius
TA-Lib==0.4.17   
